[ET-59] feat(venue) : 역참조 종속성 리팩토링 (#114)
