+++ venue/src/main/java/com/earlybird/ticket/venue/application/event/handler/VenueDeleteEventHandler.java
@@ -1,5 +1,6 @@
 package com.earlybird.ticket.venue.application.event.handler;
 
+import com.earlybird.ticket.venue.common.event.EventType;
 import com.earlybird.ticket.venue.domain.entity.Event;
 import com.earlybird.ticket.venue.application.event.dto.request.VenueDeletePayload;
 import lombok.RequiredArgsConstructor;
@@ -8,14 +9,17 @@ import org.springframework.stereotype.Component;
 @Component
 @RequiredArgsConstructor
 public class VenueDeleteEventHandler implements EventHandler<VenueDeletePayload> {
+
+    private final VenueHandlerService venueHandlerService;
+
     @Override
     public void handle(Event<VenueDeletePayload> event) {
-        //consumer 로직 작성
-        //venueService.delete();
+        VenueDeletePayload payload = event.getPayload();
+        venueHandlerService.delete(payload);
     }
 
     @Override
     public boolean support(Event<VenueDeletePayload> event) {
-        return false;
+        return event.getEventType().equals(EventType.VENUE_DELETE);
     }
 }
