+++ venue/src/main/java/com/earlybird/ticket/venue/application/event/handler/VenueHandlerService.java
@@ -1,4 +1,4 @@
-package com.earlybird.ticket.venue.application.service;
+package com.earlybird.ticket.venue.application.event.handler;
 
 import com.earlybird.ticket.venue.application.event.dto.request.VenueCreatePayload;
 import com.earlybird.ticket.venue.application.event.dto.request.VenueDeletePayload;
@@ -6,7 +6,7 @@ import com.earlybird.ticket.venue.application.event.dto.request.VenueUpdatePaylo
 import org.springframework.stereotype.Service;
 
 @Service
-public interface VenueService {
+public interface VenueHandlerService {
 
     void create(VenueCreatePayload venueCreatePayload);
